chore: refactor managementconflict logic
Put it in its own package so we can better reason about it.
