+++ pkg/k8s/meta.go
@@ -22,11 +22,8 @@ import (
 
 	operatork8s "github.com/GoogleCloudPlatform/k8s-config-connector/operator/pkg/k8s"
 	corekccv1alpha1 "github.com/GoogleCloudPlatform/k8s-config-connector/pkg/apis/core/v1alpha1"
-	"github.com/GoogleCloudPlatform/k8s-config-connector/pkg/lease/leasable"
 	"github.com/GoogleCloudPlatform/k8s-config-connector/pkg/text"
 
-	tfschema "github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	"github.com/nasa9084/go-openapi"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
@@ -132,121 +129,6 @@ func RemoveAnnotation(annotation string, obj metav1.Object) {
 	}
 	delete(annotations, annotation)
 	obj.SetAnnotations(annotations)
-}
-
-func GetManagementConflictPreventionAnnotationValue(obj metav1.Object) (ManagementConflictPreventionPolicy, error) {
-	if val, ok := GetAnnotation(ManagementConflictPreventionPolicyFullyQualifiedAnnotation, obj); ok {
-		return valueToManagementConflictPreventionPolicy(val)
-	}
-	return ManagementConflictPreventionPolicyNone,
-		fmt.Errorf("attempted to get value for annotation %v, but annotation was not found", ManagementConflictPreventionPolicyFullyQualifiedAnnotation)
-}
-
-func EnsureManagementConflictPreventionAnnotationForTFBasedResource(ctx context.Context, c client.Client, obj metav1.Object, rc *corekccv1alpha1.ResourceConfig, tfResourceMap map[string]*tfschema.Resource) error {
-	ns := corev1.Namespace{}
-	if err := c.Get(ctx, types.NamespacedName{Name: obj.GetNamespace()}, &ns); err != nil {
-		return fmt.Errorf("error getting namespace %v: %w", obj.GetNamespace(), err)
-	}
-	return ValidateOrDefaultManagementConflictPreventionAnnotationForTFBasedResource(obj, &ns, rc, tfResourceMap)
-}
-
-func ValidateOrDefaultManagementConflictPreventionAnnotationForTFBasedResource(obj metav1.Object, ns *corev1.Namespace, rc *corekccv1alpha1.ResourceConfig, tfResourceMap map[string]*tfschema.Resource) error {
-	supportsLeasing, err := leasable.ResourceConfigSupportsLeasing(rc, tfResourceMap)
-	if err != nil {
-		return err
-	}
-	return validateOrDefaultManagementConflictPreventionAnnotation(obj, ns, supportsLeasing)
-}
-
-func ValidateOrDefaultManagementConflictPreventionAnnotationForDCLBasedResource(obj metav1.Object, ns *corev1.Namespace, schema *openapi.Schema) error {
-	supportsLeasing, err := leasable.DCLSchemaSupportsLeasing(schema)
-	if err != nil {
-		return err
-	}
-	return validateOrDefaultManagementConflictPreventionAnnotation(obj, ns, supportsLeasing)
-}
-
-func validateOrDefaultManagementConflictPreventionAnnotation(obj metav1.Object, ns *corev1.Namespace, supportsLeasing bool) error {
-	value, ok := GetAnnotation(ManagementConflictPreventionPolicyFullyQualifiedAnnotation, obj)
-	if ok {
-		// the value is supplied by the customer so ensure it is valid
-		policy, err := valueToManagementConflictPreventionPolicy(value)
-		if err != nil {
-			return err
-		}
-		return validateManagementConflictPolicyForResource(policy, supportsLeasing)
-	}
-	policy, err := getDefaultManagementConflictPreventAnnotationForNamespace(ns, supportsLeasing)
-	if err != nil {
-		return err
-	}
-	SetAnnotation(ManagementConflictPreventionPolicyFullyQualifiedAnnotation, string(policy), obj)
-	return nil
-}
-
-func getDefaultManagementConflictPreventAnnotationForNamespace(ns *corev1.Namespace, supportLeasing bool) (ManagementConflictPreventionPolicy,
-	error) {
-	value, ok := GetAnnotation(ManagementConflictPreventionPolicyFullyQualifiedAnnotation, ns)
-	if ok {
-		policy, err := valueToManagementConflictPreventionPolicy(value)
-		if err != nil {
-			return ManagementConflictPreventionPolicyNone, fmt.Errorf("unable to use default management conflict policy for namespace: %w", err)
-		}
-		if !isManagementConflictPolicyValidForResource(policy, supportLeasing) {
-			return ManagementConflictPreventionPolicyNone, nil
-		}
-		return policy, nil
-	}
-	// if there is no value on the namespace return the default
-	return getDefaultManagementConflictPolicyForResource(), nil
-}
-
-func isManagementConflictPolicyValidForResource(policy ManagementConflictPreventionPolicy, supportLeasing bool) bool {
-	switch policy {
-	case ManagementConflictPreventionPolicyNone:
-		return true
-	case ManagementConflictPreventionPolicyResource:
-		return supportLeasing
-	default:
-		return false
-	}
-}
-
-// getDefaultManagementConflictPolicyForResource returns the default policy for a resource.
-//
-// This value was set to default to None, due to user complaints that label leasing behavior results
-// in resources sporadically setting not Ready, and causing issues for kpt live apply for a large
-// amount of resources.
-//
-// Before the default is flipped again, the label leaser should no longer flip the Ready state to false
-// and mark the resource as updating (https://github.com/GoogleCloudPlatform/k8s-config-connector/issues/387)
-func getDefaultManagementConflictPolicyForResource() ManagementConflictPreventionPolicy {
-	return ManagementConflictPreventionPolicyNone
-}
-
-func validateManagementConflictPolicyForResource(policy ManagementConflictPreventionPolicy, supportLeasing bool) error {
-	switch policy {
-	case ManagementConflictPreventionPolicyNone:
-		return nil
-	case ManagementConflictPreventionPolicyResource:
-		if !supportLeasing {
-			return fmt.Errorf("the resource kind does not support usage of %v of '%v'",
-				ManagementConflictPreventionPolicyAnnotation, policy)
-		}
-		return nil
-	default:
-		return fmt.Errorf("unknown management conflict policy: %v", policy)
-	}
-}
-
-func valueToManagementConflictPreventionPolicy(value string) (ManagementConflictPreventionPolicy, error) {
-	for _, policy := range ManagementConflictPreventionPolicyValues {
-		if value == string(policy) {
-			return ManagementConflictPreventionPolicy(value), nil
-		}
-	}
-	return ManagementConflictPreventionPolicyNone, fmt.Errorf("invalid management conflict policy '%v', can be one of {%v}",
-		value, strings.Join(ManagementConflictPreventionPolicyValues, ", "))
 }
 
 func SetDefaultContainerAnnotation(obj metav1.Object, ns *corev1.Namespace, containers []corekccv1alpha1.Container) error {
