+++ pkg/k8s/meta_test.go
@@ -23,8 +23,6 @@ import (
 	"github.com/GoogleCloudPlatform/k8s-config-connector/pkg/test"
 	testmain "github.com/GoogleCloudPlatform/k8s-config-connector/pkg/test/main"
 	"github.com/appscode/jsonpatch"
-	tfschema "github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	"github.com/nasa9084/go-openapi"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
@@ -386,506 +384,6 @@ func TestSetDefaultContainerAnnotation(t *testing.T) {
 			}
 		})
 	}
-}
-
-func TestValidateOrDefaultManagementConflictPreventionAnnotationForTFBasedResource(t *testing.T) {
-	tests := []struct {
-		Name                                  string
-		ManagementConflictNamespaceAnnotation string
-		ManagementConflictObjectAnnotation    string
-		MetadataMappingLabels                 string
-		LabelsFieldIsMutable                  bool
-		ExpectedObjectAnnotation              string
-		ShouldSucceed                         bool
-	}{
-		{
-			Name:                                  "none policy on namespace, empty on object",
-			ManagementConflictNamespaceAnnotation: "none",
-			ManagementConflictObjectAnnotation:    "",
-			MetadataMappingLabels:                 "",
-			ExpectedObjectAnnotation:              "none",
-			ShouldSucceed:                         true,
-		},
-		{
-			Name:                                  "none policy on namespace, resource on object",
-			ManagementConflictNamespaceAnnotation: "none",
-			ManagementConflictObjectAnnotation:    "resource",
-			MetadataMappingLabels:                 "labels_field",
-			LabelsFieldIsMutable:                  true,
-			ExpectedObjectAnnotation:              "resource",
-			ShouldSucceed:                         true,
-		},
-		{
-			Name:                                  "none policy on namespace, none on object",
-			ManagementConflictNamespaceAnnotation: "none",
-			ManagementConflictObjectAnnotation:    "none",
-			MetadataMappingLabels:                 "",
-			ExpectedObjectAnnotation:              "none",
-			ShouldSucceed:                         true,
-		},
-		{
-			Name:                                  "resource policy on namespace, empty on object",
-			ManagementConflictNamespaceAnnotation: "resource",
-			ManagementConflictObjectAnnotation:    "",
-			MetadataMappingLabels:                 "labels_field",
-			LabelsFieldIsMutable:                  true,
-			ExpectedObjectAnnotation:              "resource",
-			ShouldSucceed:                         true,
-		},
-		{
-			Name:                                  "resource policy on namespace, resource on object",
-			ManagementConflictNamespaceAnnotation: "resource",
-			ManagementConflictObjectAnnotation:    "resource",
-			MetadataMappingLabels:                 "labels_field",
-			LabelsFieldIsMutable:                  true,
-			ExpectedObjectAnnotation:              "resource",
-			ShouldSucceed:                         true,
-		},
-		{
-			Name:                                  "resource policy on namespace, none on object",
-			ManagementConflictNamespaceAnnotation: "resource",
-			ManagementConflictObjectAnnotation:    "none",
-			MetadataMappingLabels:                 "labels_field",
-			LabelsFieldIsMutable:                  true,
-			ExpectedObjectAnnotation:              "none",
-			ShouldSucceed:                         true,
-		},
-		{
-			Name:                                  "resource policy on namespace with no labels support should default to none",
-			ManagementConflictNamespaceAnnotation: "resource",
-			ManagementConflictObjectAnnotation:    "",
-			MetadataMappingLabels:                 "",
-			ExpectedObjectAnnotation:              "none",
-			ShouldSucceed:                         true,
-		},
-		{
-			Name:                                  "resource policy on namespace with immutable labels should default to none",
-			ManagementConflictNamespaceAnnotation: "resource",
-			ManagementConflictObjectAnnotation:    "",
-			MetadataMappingLabels:                 "labels_field",
-			LabelsFieldIsMutable:                  false,
-			ExpectedObjectAnnotation:              "none",
-			ShouldSucceed:                         true,
-		},
-		{
-			Name:                                  "resource policy on object should require labels support",
-			ManagementConflictNamespaceAnnotation: "",
-			ManagementConflictObjectAnnotation:    "resource",
-			MetadataMappingLabels:                 "",
-			ExpectedObjectAnnotation:              "resource",
-			ShouldSucceed:                         false,
-		},
-		{
-			Name:                                  "resource policy on object should require mutable labels",
-			ManagementConflictNamespaceAnnotation: "",
-			ManagementConflictObjectAnnotation:    "resource",
-			MetadataMappingLabels:                 "labels_field",
-			LabelsFieldIsMutable:                  false,
-			ExpectedObjectAnnotation:              "resource",
-			ShouldSucceed:                         false,
-		},
-		{
-			Name:                                  "invalid policy on namespace",
-			ManagementConflictNamespaceAnnotation: "invalid",
-			ManagementConflictObjectAnnotation:    "",
-			MetadataMappingLabels:                 "",
-			ExpectedObjectAnnotation:              "",
-			ShouldSucceed:                         false,
-		},
-		{
-			Name:                                  "invalid policy on object",
-			ManagementConflictNamespaceAnnotation: "resource",
-			ManagementConflictObjectAnnotation:    "invalid",
-			MetadataMappingLabels:                 "",
-			ExpectedObjectAnnotation:              "invalid",
-			ShouldSucceed:                         false,
-		},
-		{
-			Name:                                  "no value on namespace or resource with no labels support (i.e. default behavior when the resource doesn't support labels)",
-			ManagementConflictNamespaceAnnotation: "",
-			ManagementConflictObjectAnnotation:    "",
-			MetadataMappingLabels:                 "",
-			ExpectedObjectAnnotation:              "none",
-			ShouldSucceed:                         true,
-		},
-		{
-			Name:                                  "no value on namespace or resource with immutable labels (i.e. default behavior when the resource doesn't support mutable labels)",
-			ManagementConflictNamespaceAnnotation: "",
-			ManagementConflictObjectAnnotation:    "",
-			MetadataMappingLabels:                 "",
-			LabelsFieldIsMutable:                  false,
-			ExpectedObjectAnnotation:              "none",
-			ShouldSucceed:                         true,
-		},
-		{
-			Name:                                  "no value on namespace or resource with mutable labels (i.e. default behavior when the resource supports mutable labels)",
-			ManagementConflictNamespaceAnnotation: "",
-			ManagementConflictObjectAnnotation:    "",
-			MetadataMappingLabels:                 "labels_value",
-			LabelsFieldIsMutable:                  true,
-			ExpectedObjectAnnotation:              "none",
-			ShouldSucceed:                         true,
-		},
-	}
-	for _, tc := range tests {
-		t.Run(tc.Name, func(t *testing.T) {
-			ns := corev1.Namespace{}
-			ns.SetName("my-namespace")
-			ns.SetAnnotations(newManagementConflictAnnotations(tc.ManagementConflictNamespaceAnnotation))
-			obj := unstructured.Unstructured{}
-			obj.SetAnnotations(newManagementConflictAnnotations(tc.ManagementConflictObjectAnnotation))
-
-			fakeTFResourceName := "google_fake_resource"
-			fakeTFResource := &tfschema.Resource{
-				Schema: map[string]*tfschema.Schema{},
-			}
-			fakeTFLabelsField := tc.MetadataMappingLabels
-			if fakeTFLabelsField != "" {
-				fakeTFResource.Schema[fakeTFLabelsField] = &tfschema.Schema{
-					ForceNew: !tc.LabelsFieldIsMutable,
-				}
-			}
-			fakeTFProvider := &tfschema.Provider{
-				ResourcesMap: map[string]*tfschema.Resource{
-					fakeTFResourceName: fakeTFResource,
-				},
-			}
-			rc := corekccv1alpha1.ResourceConfig{
-				Name: fakeTFResourceName,
-				MetadataMapping: corekccv1alpha1.MetadataMapping{
-					Labels: fakeTFLabelsField,
-				},
-			}
-
-			err := k8s.ValidateOrDefaultManagementConflictPreventionAnnotationForTFBasedResource(&obj, &ns, &rc, fakeTFProvider.ResourcesMap)
-			if tc.ShouldSucceed != (err == nil) {
-				t.Fatalf("expected success to be '%v', instead got error mismsatch: %v", tc.ShouldSucceed, err)
-			}
-			value, ok := k8s.GetAnnotation(k8s.ManagementConflictPreventionPolicyFullyQualifiedAnnotation, &obj)
-			if ok || tc.ExpectedObjectAnnotation != "" {
-				if value != tc.ExpectedObjectAnnotation {
-					t.Fatalf("unexpected management conflict annotation value: got '%v', want '%v'", value, tc.ExpectedObjectAnnotation)
-				}
-			}
-		})
-	}
-}
-
-func TestValidateOrDefaultManagementConflictPreventionAnnotationForDCLBasedResource(t *testing.T) {
-	tests := []struct {
-		Name                                  string
-		ManagementConflictNamespaceAnnotation string
-		ManagementConflictObjectAnnotation    string
-		Schema                                *openapi.Schema
-		ExpectedObjectAnnotation              string
-		ShouldSucceed                         bool
-	}{
-		{
-			Name:                                  "none policy on namespace, empty on object",
-			ManagementConflictNamespaceAnnotation: "none",
-			ManagementConflictObjectAnnotation:    "",
-			Schema: &openapi.Schema{
-				Type: "object",
-				Properties: map[string]*openapi.Schema{
-					"labels": &openapi.Schema{
-						Type: "string",
-					},
-				},
-				Extension: map[string]interface{}{
-					"x-dcl-labels": "labels",
-				},
-			},
-			ExpectedObjectAnnotation: "none",
-			ShouldSucceed:            true,
-		},
-		{
-			Name:                                  "none policy on namespace, resource on object",
-			ManagementConflictNamespaceAnnotation: "none",
-			ManagementConflictObjectAnnotation:    "resource",
-			Schema: &openapi.Schema{
-				Type: "object",
-				Properties: map[string]*openapi.Schema{
-					"labels": &openapi.Schema{
-						Type: "string",
-					},
-				},
-				Extension: map[string]interface{}{
-					"x-dcl-labels": "labels",
-				},
-			},
-			ExpectedObjectAnnotation: "resource",
-			ShouldSucceed:            true,
-		},
-		{
-			Name:                                  "none policy on namespace, none on object",
-			ManagementConflictNamespaceAnnotation: "none",
-			ManagementConflictObjectAnnotation:    "none",
-			Schema: &openapi.Schema{
-				Type: "object",
-			},
-			ExpectedObjectAnnotation: "none",
-			ShouldSucceed:            true,
-		},
-		{
-			Name:                                  "resource policy on namespace, empty on object",
-			ManagementConflictNamespaceAnnotation: "resource",
-			ManagementConflictObjectAnnotation:    "",
-			Schema: &openapi.Schema{
-				Type: "object",
-				Properties: map[string]*openapi.Schema{
-					"labels": &openapi.Schema{
-						Type: "string",
-					},
-				},
-				Extension: map[string]interface{}{
-					"x-dcl-labels": "labels",
-				},
-			},
-			ExpectedObjectAnnotation: "resource",
-			ShouldSucceed:            true,
-		},
-		{
-			Name:                                  "resource policy on namespace, resource on object",
-			ManagementConflictNamespaceAnnotation: "resource",
-			ManagementConflictObjectAnnotation:    "resource",
-			Schema: &openapi.Schema{
-				Type: "object",
-				Properties: map[string]*openapi.Schema{
-					"labels": &openapi.Schema{
-						Type: "string",
-					},
-				},
-				Extension: map[string]interface{}{
-					"x-dcl-labels": "labels",
-				},
-			},
-			ExpectedObjectAnnotation: "resource",
-			ShouldSucceed:            true,
-		},
-		{
-			Name:                                  "resource policy on namespace, none on object",
-			ManagementConflictNamespaceAnnotation: "resource",
-			ManagementConflictObjectAnnotation:    "none",
-			Schema: &openapi.Schema{
-				Type: "object",
-				Properties: map[string]*openapi.Schema{
-					"labels": &openapi.Schema{
-						Type: "string",
-					},
-				},
-				Extension: map[string]interface{}{
-					"x-dcl-labels": "labels",
-				},
-			},
-			ExpectedObjectAnnotation: "none",
-			ShouldSucceed:            true,
-		},
-		{
-			Name:                                  "resource policy on namespace with no labels support should default to none",
-			ManagementConflictNamespaceAnnotation: "resource",
-			Schema: &openapi.Schema{
-				Type: "object",
-			},
-			ExpectedObjectAnnotation: "none",
-			ShouldSucceed:            true,
-		},
-		{
-			Name:                                  "resource policy on namespace with immutable labels should default to none",
-			ManagementConflictNamespaceAnnotation: "resource",
-			ManagementConflictObjectAnnotation:    "",
-			Schema: &openapi.Schema{
-				Type: "object",
-				Properties: map[string]*openapi.Schema{
-					"labels": &openapi.Schema{
-						Type: "string",
-						Extension: map[string]interface{}{
-							"x-kubernetes-immutable": true,
-						},
-					},
-				},
-				Extension: map[string]interface{}{
-					"x-dcl-labels": "labels",
-				},
-			},
-			ExpectedObjectAnnotation: "none",
-			ShouldSucceed:            true,
-		},
-		{
-			Name:                                  "resource policy on object should require labels support",
-			ManagementConflictNamespaceAnnotation: "",
-			ManagementConflictObjectAnnotation:    "resource",
-			Schema: &openapi.Schema{
-				Type: "object",
-			},
-			ExpectedObjectAnnotation: "resource",
-			ShouldSucceed:            false,
-		},
-		{
-			Name:                                  "resource policy on object should require mutable labels",
-			ManagementConflictNamespaceAnnotation: "",
-			ManagementConflictObjectAnnotation:    "resource",
-			Schema: &openapi.Schema{
-				Type: "object",
-				Properties: map[string]*openapi.Schema{
-					"labels": &openapi.Schema{
-						Type: "string",
-						Extension: map[string]interface{}{
-							"x-kubernetes-immutable": true,
-						},
-					},
-				},
-				Extension: map[string]interface{}{
-					"x-dcl-labels": "labels",
-				},
-			},
-			ExpectedObjectAnnotation: "resource",
-			ShouldSucceed:            false,
-		},
-		{
-			Name:                                  "invalid policy on namespace",
-			ManagementConflictNamespaceAnnotation: "invalid",
-			ManagementConflictObjectAnnotation:    "",
-			Schema: &openapi.Schema{
-				Type: "object",
-			},
-			ExpectedObjectAnnotation: "",
-			ShouldSucceed:            false,
-		},
-		{
-			Name:                                  "invalid policy on object",
-			ManagementConflictNamespaceAnnotation: "resource",
-			ManagementConflictObjectAnnotation:    "invalid",
-			Schema: &openapi.Schema{
-				Type: "object",
-			},
-			ExpectedObjectAnnotation: "invalid",
-			ShouldSucceed:            false,
-		},
-		{
-			Name:                                  "no value on namespace or resource with no labels support (i.e. default behavior when the resource doesn't support labels)",
-			ManagementConflictNamespaceAnnotation: "",
-			ManagementConflictObjectAnnotation:    "",
-			Schema: &openapi.Schema{
-				Type: "object",
-			},
-			ExpectedObjectAnnotation: "none",
-			ShouldSucceed:            true,
-		},
-		{
-			Name:                                  "no value on namespace or resource with immutable labels (i.e. default behavior when the resource doesn't support mutable labels)",
-			ManagementConflictNamespaceAnnotation: "",
-			ManagementConflictObjectAnnotation:    "",
-			Schema: &openapi.Schema{
-				Type: "object",
-				Properties: map[string]*openapi.Schema{
-					"labels": &openapi.Schema{
-						Type: "string",
-						Extension: map[string]interface{}{
-							"x-kubernetes-immutable": true,
-						},
-					},
-				},
-				Extension: map[string]interface{}{
-					"x-dcl-labels": "labels",
-				},
-			},
-			ExpectedObjectAnnotation: "none",
-			ShouldSucceed:            true,
-		},
-		{
-			Name:                                  "no value on namespace or resource with mutable labels (i.e. default behavior when the resource supports mutable labels)",
-			ManagementConflictNamespaceAnnotation: "",
-			ManagementConflictObjectAnnotation:    "",
-			Schema: &openapi.Schema{
-				Type: "object",
-				Properties: map[string]*openapi.Schema{
-					"labels": &openapi.Schema{
-						Type: "string",
-					},
-				},
-				Extension: map[string]interface{}{
-					"x-dcl-labels": "labels",
-				},
-			},
-			ExpectedObjectAnnotation: "none",
-			ShouldSucceed:            true,
-		},
-	}
-	for _, tc := range tests {
-		t.Run(tc.Name, func(t *testing.T) {
-			ns := corev1.Namespace{}
-			ns.SetName("my-namespace")
-			ns.SetAnnotations(newManagementConflictAnnotations(tc.ManagementConflictNamespaceAnnotation))
-			obj := unstructured.Unstructured{}
-			obj.SetAnnotations(newManagementConflictAnnotations(tc.ManagementConflictObjectAnnotation))
-
-			err := k8s.ValidateOrDefaultManagementConflictPreventionAnnotationForDCLBasedResource(&obj, &ns, tc.Schema)
-			if tc.ShouldSucceed != (err == nil) {
-				t.Fatalf("expected success to be '%v', instead got error mismsatch: %v", tc.ShouldSucceed, err)
-			}
-			value, ok := k8s.GetAnnotation(k8s.ManagementConflictPreventionPolicyFullyQualifiedAnnotation, &obj)
-			if ok || tc.ExpectedObjectAnnotation != "" {
-				if value != tc.ExpectedObjectAnnotation {
-					t.Fatalf("unexpected management conflict annotation value: got '%v', want '%v'", value, tc.ExpectedObjectAnnotation)
-				}
-			}
-		})
-	}
-}
-
-func TestGetManagementConflictPreventionAnnotationValue(t *testing.T) {
-	testCases := []struct {
-		Name           string
-		Annotations    map[string]string
-		ExpectedPolicy k8s.ManagementConflictPreventionPolicy
-		ShouldSucceed  bool
-	}{
-		{
-			Name:           "nil annotations should error",
-			Annotations:    nil,
-			ExpectedPolicy: k8s.ManagementConflictPreventionPolicyNone,
-			ShouldSucceed:  false,
-		},
-		{
-			Name:           "missing annotation should error",
-			Annotations:    make(map[string]string),
-			ExpectedPolicy: k8s.ManagementConflictPreventionPolicyNone,
-			ShouldSucceed:  false,
-		},
-		{
-			Name:           "invalid annotation should error",
-			Annotations:    newManagementConflictAnnotations("my invalid policy name"),
-			ExpectedPolicy: k8s.ManagementConflictPreventionPolicyNone,
-			ShouldSucceed:  false,
-		},
-		{
-			Name:           "valid value should succeed",
-			Annotations:    newManagementConflictAnnotations(k8s.ManagementConflictPreventionPolicyResource),
-			ExpectedPolicy: k8s.ManagementConflictPreventionPolicyResource,
-			ShouldSucceed:  true,
-		},
-	}
-	for _, tc := range testCases {
-		t.Run(tc.Name, func(t *testing.T) {
-			obj := unstructured.Unstructured{}
-			obj.SetAnnotations(tc.Annotations)
-			policy, err := k8s.GetManagementConflictPreventionAnnotationValue(&obj)
-			if tc.ShouldSucceed != (err == nil) {
-				t.Fatalf("expected success to be '%v', instead got error mismatch: %v", tc.ShouldSucceed, err)
-			}
-			if policy != tc.ExpectedPolicy {
-				t.Fatalf("policy mismatch: got '%v', want '%v'", policy, tc.ExpectedPolicy)
-			}
-		})
-	}
-}
-
-func newManagementConflictAnnotations(policy string) map[string]string {
-	annotations := make(map[string]string)
-	if policy != "" {
-		annotations[k8s.ManagementConflictPreventionPolicyFullyQualifiedAnnotation] = policy
-	}
-	return annotations
 }
 
 func TestIsManagedByKCC(t *testing.T) {
